refactor(webui): avoid importing `expo-router` in API routes
